FINI

+++ ProjetBanque/socket/central.cpp
@@ -199,11 +199,11 @@ string get_all_account() {
     int banq;
 
     boost::split(result, data, boost::is_any_of("$"));
-    for (size_t i = 0; i < result.size(); i++)
+    for (int i = 0; i < result.size(); i++)
     {
         banq = i+1;
         boost::split(res, result[i], boost::is_any_of("-"));
-        for (size_t i = 0; i < res[i].size(); i++)
+        for (int i = 0; i < res[i].size(); i++)
         {
             pta.put("Nombre", res[i]);
             pta.put("Banque", std::to_string(banq));
